Set DXL Workbench tool when hardware error exists
The workbench is designed very very well, so that it only works when a
device was successfully pinged. So when we receive a ping answer that
indicates a hardware error (error byte = 0x80), we only get a failed
ping attempt but can't do any more reading/writing/rebooting of the
device.
Now, when the ping returns with the error byte indicating hardware
error, it will still return a failed ping but at least initializes the
tool for further communication.